test(server): relocate db story test, remove tests-database crate
