--- crates/tests-database/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod helpers;
